finished tests, now with data validation updated the README for how to use the API calls

+++ graphserver.py
@@ -32,29 +32,52 @@ def show_entries():
 
 @app.route('/branches', methods=['GET', 'POST'])
 def branches():
-    # TODO - form validation
     if request.method == 'POST':
         if request.form['_method'] == "delete":
-            results = app.con.execute(app.branches.delete().where(app.branches.c.id == request.form['id']))
-            flash('Branch %s was successfully deleted' % request.form['branch_name'])
+            exists = app.con.execute(app.branches.select().where(app.branches.c.id == request.form['id']))
+            if exists.returns_rows:
+                results = app.con.execute(app.branches.delete().where(app.branches.c.id == request.form['id']))
+                flash('Branch %s was successfully deleted' % request.form['branch_name'])
         elif request.form['_method'] == "insert":
-            results = app.con.execute(app.branches.insert(), name=request.form['branch_name'])
-            flash('New branch %s was successfully added' % request.form['branch_name'])
-    if request.method == 'GET' and request.args.get('format') == 'json':
+            exists = app.con.execute(app.branches.select().where(app.branches.c.name == request.form['branch_name']))
+            if exists.fetchone() != None:
+                flash('Branch name exists, please enter a unique name')
+            elif request.form['branch_name'] == '':
+                flash('Please enter a branch name')
+            else:
+                results = app.con.execute(app.branches.insert(), name=request.form['branch_name'])
+                flash('New branch %s was successfully added' % request.form['branch_name'])
+    if request.method == 'GET' and request.args.get('format') == 'json' or 'application/json' in request.headers.get('Accept', ''):
         return jsonify(app.branches.select().execute().fetchall())
-    if 'application/json' in request.headers.get('Accept', '') or request.form.get('format') == 'json':
-        return jsonify(results)
     return redirect(url_for('show_entries'))
 
 @app.route('/machines', methods=['GET', 'POST'])
 def machines():
-    # TODO - form validation
     if request.method == 'POST':
         if request.form['_method'] == "delete":
-            results = app.con.execute(app.machines.delete().where(app.machines.c.id == request.form['id']))
-            flash("Machine '%s' was successfully deleted" % request.form['machine_name'])
+            exists = app.con.execute(app.machines.select().where(app.machines.c.id == request.form['id']))
+            if exists.returns_rows:
+                results = app.con.execute(app.machines.delete().where(app.machines.c.id == request.form['id']))
+                flash("Machine '%s' was successfully deleted" % request.form['machine_name'])
         elif request.form['_method'] == "insert":
-            results = app.con.execute(
+            errors = False
+            for key,value in request.form.items():
+                if key == 'machine_name':
+                    exists = app.con.execute(app.machines.select().where(app.machines.c.name == request.form['machine_name']))
+                    if exists.fetchone() != None:
+                        flash('Machine name exists, please enter a unique name')
+                        errors = True
+                    elif value == '':
+                        flash('Please provide a machine name')
+                        errors = True
+                elif key in ('os_id', 'is_throttling', 'cpu_speed', 'is_active'):
+                    try:
+                        i = float(value)
+                    except ValueError, TypeError:
+                        flash('%s must have a numeric value' % key)
+                        errors = True
+            if not errors:
+                results = app.con.execute(
                             app.machines.insert(), 
                             os_id=int(request.form['os_id']),
                             is_throttling=int(request.form['is_throttling']),
@@ -63,15 +86,13 @@ def machines():
                             is_active=int(request.form['is_active']), 
                             date_added=int(time.time())
                         )
-            flash('New machine %s was successfully added' % request.form['machine_name'])
-    if request.method == 'GET' and request.args.get('format') == 'json':
+                flash('New machine %s was successfully added' % request.form['machine_name'])
+    if request.method == 'GET' and request.args.get('format') == 'json' or 'application/json' in request.headers.get('Accept', ''):
         machines = {}
         results = app.machines.select().execute().fetchall()
         for r in results:
             machines[r[0]] = r[4]
         return jsonify(machines)
-    if 'application/json' in request.headers.get('Accept', '') or request.form.get('format') == 'json':
-        return jsonify(results)
     return redirect(url_for('show_entries'))
 
 if __name__ == '__main__':
